TipModal: Move "Buy More" next to price selection.
This change is primarily to make space for 4492 `Make support (tip case) dialog clearer`. The bottom section should be reserved for elements that will finalize the action of the modal.

Also, having the "Buy More LBC" at the bottom seems out of place, since the context of LBC/amount is in the middle of the modal.

Cleared lint for unused `isMobile`.